Lift almide-frontend to codopsy grade A
