+++ crates/almide-frontend/src/check/mod.rs
@@ -22,6 +22,7 @@
 mod types;
 mod infer;
 pub(crate) mod calls;
+mod retired_aliases;
 mod builtin_calls;
 mod static_dispatch;
 mod solving;
